--- crates/rspack_plugin_manifest/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-edition    = "2021"
-license    = "MIT"
-name       = "rspack_plugin_manifest"
-repository = "https://github.com/web-infra-dev/rspack"
-version    = "0.1.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dev-dependencies]
-rspack_testing = { path = "../rspack_testing" }
-testing_macros = { workspace = true }
-
-[dependencies]
-async-recursion = { workspace = true }
-async-trait     = { workspace = true }
-rspack_core     = { path = "../rspack_core" }
-rspack_error    = { path = "../rspack_error" }
-rspack_regex    = { path = "../rspack_regex" }
