+++ webpack-test/configCases/plugins/lib-manifest-plugin/test.filter.js
