+++ webpack-test/configCases/plugins/lib-manifest-plugin/webpack.config.js
@@ -1,46 +1,18 @@
 var path = require("path");
-// var LibManifestPlugin = require("../../../../").LibManifestPlugin;
-const manifestPlugin = require("rspack-manifest-plugin").WebpackManifestPlugin;
+var LibManifestPlugin = require("../../../../").LibManifestPlugin;
 
 /** @type {function(any, any): import("@rspack/core").Configuration} */
-module.exports = (env, { testPath }) => {
-	console.log(testPath)
-return ({
+module.exports = (env, { testPath }) => ({
 	entry: {
-		bundle0: ["./"],
+		bundle0: ["./"]
 	},
 	plugins: [
-		new manifestPlugin({
-			fileName: path.resolve(testPath, "[name]-manifest.json"),
-			generate: (seed, files, entries) => {
-					console.log(files)
-				const manifestFiles = files.reduce((manifest, file) => {
-					manifest[file.name] = file.path;
-					return manifest;
-				}, seed);
-				const entrypointFiles = Object.keys(entries).reduce(
-					(previous, name) =>
-						previous.concat(
-							entries[name].filter((fileName) => !fileName.endsWith(".map")),
-						),
-					[],
-				);
-				return {
-					files: manifestFiles,
-					entrypoints: entrypointFiles,
-				};
-			},
-		}),
-		// new LibManifestPlugin({
-		// 	path: path.resolve(testPath, "[name]-manifest.json"),
-		// 	name: "[name]_[fullhash]"
-		// })
+		new LibManifestPlugin({
+			path: path.resolve(testPath, "[name]-manifest.json"),
+			name: "[name]_[fullhash]"
+		})
 	],
-	stats: {
-		all: false,
-	},
 	node: {
-		__dirname: false,
-	},
-})
-};
+		__dirname: false
+	}
+});
